реализованы все методы для пользователей

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -3,6 +3,7 @@ package ru.practicum.shareit.user;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.exceptions.EmailAlreadyExistsException;
+import ru.practicum.shareit.exceptions.UserNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,21 +28,29 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(Long id, User user) {
+        if (userRepository.getUser(id) == null) {
+            throw new UserNotFoundException();
+        }
+        if (userRepository.existsByEmail(user.getEmail())) {
+            throw new EmailAlreadyExistsException();
+        }
         return userRepository.updateUser(id, user);
     }
 
     @Override
     public User getUser(Long id) {
-        return userRepository.getUser(id);
-    }
-
-    @Override
-    public List<User> getAllUsers() {
-        return userRepository.getAllUsers();
+        User user = userRepository.getUser(id);
+        if (user == null) {
+            throw new UserNotFoundException();
+        }
+        return user;
     }
 
     @Override
     public boolean deleteUser(Long id) {
+        if (userRepository.getUser(id) == null) {
+            throw new UserNotFoundException();
+        }
         return userRepository.deleteUser(id);
     }
 }
